+++ src/app/hooks/use-geometry-from-duckdb.ts
@@ -5,6 +5,8 @@ import {
 	MaterialData,
 	TransformData,
 	GeometryInstance,
+} from "@/app/utils/types";
+import {
 	createThreeMaterial,
 	createTransformMatrix,
 	createBufferGeometryFromMesh,
