+++ src/app/utils/types.ts
@@ -5,6 +5,7 @@ import { useQueryObjects } from "../hooks/use-query-objects";
 import useQueryViewerAndEditor from "../hooks/use-query-viewer-and-editor";
 import { useRunDuckDbQuery } from "../hooks/use-run-duckdb-query";
 import * as duckdb from "@duckdb/duckdb-wasm";
+import * as THREE from "three";
 
 export const validFileNames = [
 	"Descriptors.parquet",
@@ -221,3 +222,109 @@ export const analyticReadinessTitles = Object.freeze([
 ] as const);
 export type AnalyticsReadinessLevels = (typeof analyticReadinessLevels)[number];
 export type AnalyticsReadinessTitle = (typeof analyticReadinessTitles)[number];
+
+// for geometrical data processing
+export type GeometryCacheData = {
+	vertices: VertexData[];
+	indices: IndexData[];
+	meshes: MeshData[];
+	materials: MaterialData[];
+	transforms: TransformData[];
+	instanceData: InstanceData[];
+};
+
+export type InstanceData = {
+	instance_index: number;
+	entity_index: number;
+	material_index: number;
+	mesh_index: number;
+	transform_index: number;
+	vertex_offset: number;
+	index_offset: number;
+	tx: number;
+	ty: number;
+	tz: number;
+	qx: number;
+	qy: number;
+	qz: number;
+	qw: number;
+	sx: number;
+	sy: number;
+	sz: number;
+	red: number;
+	green: number;
+	blue: number;
+	alpha: number;
+	roughness: number;
+	metallic: number;
+};
+
+export type FilteredGeometryResult = {
+	scene: THREE.Group | null;
+	instanceCount: number;
+	totalCount: number;
+};
+
+export type GeometryContextValue = {
+	loading: boolean;
+	error: Error | null;
+	getFilteredScene: (entityIndices: number[]) => FilteredGeometryResult;
+};
+
+export type VertexData = {
+	index: number;
+	x: number;
+	y: number;
+	z: number;
+};
+
+export type IndexData = {
+	index: number;
+	index_value: number;
+};
+
+export type MeshData = {
+	index: number;
+	vertex_offset: number;
+	index_offset: number;
+};
+
+export type MaterialData = {
+	index: number;
+	red: number;
+	green: number;
+	blue: number;
+	alpha: number;
+	roughness: number;
+	metallic: number;
+};
+
+export type TransformData = {
+	index: number;
+	tx: number;
+	ty: number;
+	tz: number;
+	qx: number;
+	qy: number;
+	qz: number;
+	qw: number;
+	sx: number;
+	sy: number;
+	sz: number;
+};
+
+export type GeometryInstance = {
+	instanceIndex: number;
+	entityIndex: number;
+	meshIndex: number;
+	transformIndex: number;
+	materialIndex: number;
+	LocalId: string;
+	GlobalId: string;
+	entityName: string;
+	category: string;
+	vertexOffset: number;
+	indexOffset: number;
+	transform: THREE.Matrix4;
+	material: THREE.MeshStandardMaterial;
+};
